EchoLearn.Api.Tests/Services/SpeechServiceTests.cs
using Moq;
using Microsoft.Extensions.Logging;
using Microsoft.Extensions.Configuration;
using Microsoft.CognitiveServices.Speech;
using Microsoft.CognitiveServices.Speech.Audio;
using EchoLearn.Api.Services;

namespace EchoLearnAI.Api.Tests.Services
{
    public class SpeechServiceTests
    {
        private readonly Mock<ILogger<SpeechService>> _loggerMock;
        private readonly Mock<IConfiguration> _configurationMock;
        private readonly Mock<ISpeechRecognizerFactory> _speechRecognizerFactoryMock;
        private readonly SpeechService _speechService;

        public SpeechServiceTests()
        {
            _loggerMock = new Mock<ILogger<SpeechService>>();
            _configurationMock = new Mock<IConfiguration>();
            _speechRecognizerFactoryMock = new Mock<ISpeechRecognizerFactory>();

            // Set up configuration mock
            _configurationMock.Setup(config => config["AzureCognitiveServices:Speech:SubscriptionKey"])
                              .Returns("FakeSubscriptionKey");
            _configurationMock.Setup(config => config["AzureCognitiveServices:Speech:Region"])
                              .Returns("FakeRegion");

            _speechService = new SpeechService(
                _loggerMock.Object,
                _configurationMock.Object,
                _speechRecognizerFactoryMock.Object);
        }
/*
        [Fact]
        public async Task RecognizeSpeechAsync_ReturnsRecognizedText_WhenSpeechIsRecognized()
        {
            // Arrange
            using var audioStream = new MemoryStream(new byte[] { 1, 2, 3 });

            var mockResult = new SpeechRecognitionResult
            {
                Reason = ResultReason.RecognizedSpeech,
                Text = "Hello world"
            };

            var speechRecognizerMock = new Mock<SpeechRecognizer>();
            speechRecognizerMock.Setup(r => r.RecognizeOnceAsync())
                                .ReturnsAsync(mockResult);

            _speechRecognizerFactoryMock.Setup(f => f.CreateSpeechRecognizer(It.IsAny<SpeechConfig>(), It.IsAny<AudioConfig>()))
                                        .Returns(speechRecognizerMock.Object);

            // Act
            var result = await _speechService.RecognizeSpeechAsync(audioStream);

            // Assert
            Assert.Equal("Hello world", result);
        }

        [Fact]
        public async Task RecognizeSpeechAsync_ReturnsErrorMessage_WhenSpeechRecognitionFails()
        {
            // Arrange
            using var audioStream = new MemoryStream(new byte[] { 1, 2, 3 });

            var mockResult = new SpeechRecognitionResult
            {
                Reason = ResultReason.NoMatch,
                Text = string.Empty
            };

            var speechRecognizerMock = new Mock<SpeechRecognizer>();
            speechRecognizerMock.Setup(r => r.RecognizeOnceAsync())
                                .ReturnsAsync(mockResult);

            _speechRecognizerFactoryMock.Setup(f => f.CreateSpeechRecognizer(It.IsAny<SpeechConfig>(), It.IsAny<AudioConfig>()))
                                        .Returns(speechRecognizerMock.Object);

            // Act
            var result = await _speechService.RecognizeSpeechAsync(audioStream);

            // Assert
            Assert.Contains("Speech recognition failed", result);
        }

        [Fact]
        public async Task RecognizeSpeechAsync_ReturnsErrorMessage_WhenExceptionOccurs()
        {
            // Arrange
            Stream audioStream = null;

            // Act
            var result = await _speechService.RecognizeSpeechAsync(audioStream);

            // Assert
            Assert.Equal("An error occurred during speech recognition.", result);
        }*/
    }
}
